requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
python-dotenv
Flask-WTF
wtforms
email-validator